samples/python/azureai-streaming-poem-agent/requirements.txt
python-dotenv
aiohttp
openai
azure-identity
microsoft-agents-activity
microsoft-agents-hosting-core
microsoft-agents-hosting-aiohttp
microsoft-agents-authentication-msal